requirements.txt
requests
feedparser
python-dotenv
newspaper3k
lxml_html_clean